+++ package.json
@@ -21,7 +21,8 @@
   "files": [
     "dist",
     "src",
-    "bin"
+    "bin",
+    "tsconfig.json"
   ],
   "author": "Flare Networks",
   "homepage": "https://github.com/flare-foundation/p-chain-staking-code#readme",
@@ -29,12 +30,11 @@
   "devDependencies": {
     "@types/commander": "^2.12.2",
     "@types/elliptic": "^6.4.14",
+    "@types/ethereumjs-util": "^6.1.0",
     "@types/node": "*",
+    "@types/node-fetch": "^2.6.4",
     "@types/node-hid": "^1.3.1",
     "@types/ws": "^8.5.3",
-    "@types/ethereumjs-util": "^6.1.0",
-    "ethereumjs-util": "^7.1.5",
-    "fireblocks-sdk": "^5.1.4",
     "ts-node": "^10.9.1",
     "typescript": "^4.8.4"
   },
@@ -45,8 +45,8 @@
     "bech32": "^2.0.0",
     "chalk": "4.1.2",
     "commander": "^9.4.1",
+    "ethereumjs-util": "^7.1.5",
     "dotenv": "^16.0.1",
-    "ethereumjs-tx": "^2.1.2",
     "ethers": "6.5.1",
     "web3": "^1.8.0"
   }
