+++ src/forDefi/forDefi.ts
@@ -1,9 +1,10 @@
 
-const fetch = require('node-fetch')
+import fetch from 'node-fetch'
 import { readFileSync, writeFileSync } from 'fs'
 import crypto from "crypto"
-import { unPrefix0x, readUnsignedTxJson, readUnsignedWithdrawalTx } from "../src/utils"
-import { SignedTxJson, SignedWithdrawalTxJson, UnsignedTxJson, UnsignedWithdrawalTxJson, ContextFile } from './interfaces'
+import { readUnsignedWithdrawalTx } from "./utils"
+import { unPrefix0x, readUnsignedTxJson } from "../utils"
+import { SignedTxJson, SignedWithdrawalTxJson, UnsignedTxJson, UnsignedWithdrawalTxJson, ContextFile } from '../interfaces'
 
 const gatewayHost = "api.fordefi.com"
 
@@ -61,7 +62,7 @@ export async function sendToForDefi(unsignedTxidFile: string, ctxFile: string, w
         headers: {
             'Content-Type': 'application/json',
             "Authorization": `Bearer ${accessToken}`,
-            'X-Timestamp': timestamp,
+            'X-Timestamp': timestamp.toString(),
             'X-Signature': signature1,
         },
         body: requestBody,
